+++ views/category.css.php
@@ -38,18 +38,16 @@ if ( empty( $this->terms ) && ! empty( $options['terms'] ) ) {
 	<?php Extras::fix_category_link_css( $slug ); ?>
 	<?php Extras::add_map_link_css( $slug ); ?>
 	<?php Extras::add_week_link_css( $slug ); ?>
+	<?php Extras::add_list_link_css( $slug ); ?>
 	<?php Extras::override_customizer( $slug ); ?>
-	<?php Extras::add_deprecated_link_css( $slug ); ?>
-	<?php if ( '1' === $options['color_widgets'] ) : ?>
-		<?php Widgets::add_widget_link_css( $slug ); ?>
-	<?php endif ?>
+	<?php //Extras::add_deprecated_link_css( $slug ); ?>
+	<?php Widgets::add_widget_link_css( $slug ); ?>
 #tribe-events-content table.tribe-events-calendar .tribe-event-featured.tribe-events-category-<?php echo $slug; ?> .tribe-events-month-event-title a,
 .teccc-legend .tribe-events-category-<?php echo $slug; ?> a,
 .tribe-events-calendar .tribe-events-category-<?php echo $slug; ?> a,
 #tribe-events-content .teccc-legend .tribe-events-category-<?php echo $slug; ?> a,
 #tribe-events-content .tribe-events-calendar .tribe-events-category-<?php echo $slug; ?> a,
 .tribe-events-category-<?php echo $slug; ?> > div.hentry.vevent > h3.entry-title a,
-.tribe-events-category-<?php echo $slug; ?> h3.tribe-events-list-event-title a,
 .tribe-events-mobile.tribe-events-category-<?php echo $slug; ?> h4 a {
 	<?php if ( 'no_color' !== $options[ "{$slug}-text" ] ) : ?>
 	color: <?php esc_attr_e( $options[ "{$slug}-text" ] ); ?>;
@@ -60,19 +58,17 @@ if ( empty( $this->terms ) && ! empty( $options['terms'] ) ) {
 	<?php Extras::fix_category_background_css( $slug ); ?>
 	<?php Extras::add_map_background_css( $slug ); ?>
 	<?php Extras::add_week_background_css( $slug ); ?>
-	<?php Extras::add_deprecated_week_background_css( $slug ); ?>
-	<?php Extras::add_deprecated_background_css( $slug ); ?>
+	<?php Extras::add_list_background_css( $slug ); ?>
+	<?php //Extras::add_deprecated_week_background_css( $slug ); ?>
+	<?php //Extras::add_deprecated_background_css( $slug ); ?>
 	<?php Extras::override_customizer( $slug ); ?>
-	<?php if ( '1' === $options['color_widgets'] ) : ?>
-		<?php Widgets::add_widget_background_css( $slug ); ?>
-<?php endif ?>
+	<?php Widgets::add_widget_background_css( $slug ); ?>
 .events-archive.events-gridview #tribe-events-content table .type-tribe_events.tribe-events-category-<?php echo $slug; ?>,
-.tribe-events-category-<?php echo $slug; ?> h2.tribe-events-list-event-title.entry-title a,
+.type-tribe_events.tribe-events-category-<?php echo $slug; ?> h2,
 .teccc-legend .tribe-events-category-<?php echo $slug; ?>,
 .tribe-events-calendar .tribe-events-category-<?php echo $slug; ?>,
 #tribe-events-content .tribe-events-category-<?php echo $slug; ?> > .tribe-events-tooltip h3,
 .tribe-events-category-<?php echo $slug; ?> > div.hentry.vevent > h3.entry-title,
-.tribe-events-category-<?php echo $slug; ?> h3.tribe-events-list-event-title,
 .tribe-events-mobile.tribe-events-category-<?php echo $slug; ?> h4 {
 	<?php if ( null !== $options[ "{$slug}-background" ] ) : ?>
 	background-color: <?php esc_attr_e( $options[ "{$slug}-background" ] ); ?>;
@@ -92,11 +88,9 @@ if ( empty( $this->terms ) && ! empty( $options['terms'] ) ) {
 	<?php Extras::fix_transparent_week_background( $slug ); ?>
 
 	<?php Extras::add_map_link_css( $slug ); ?>
-	<?php if ( '1' === $options['color_widgets'] ) : ?>
-		<?php Widgets::add_widget_link_css( $slug ); ?>
-<?php endif ?>
-.tribe-events-category-<?php echo $slug; ?> h2.tribe-events-list-event-title.entry-title a,
-.tribe-events-category-<?php echo $slug; ?> h2.tribe-events-list-event-title a {
+	<?php Widgets::add_widget_link_css( $slug ); ?>
+	<?php Extras::add_list_link_css( $slug ); ?>
+{
 	width: auto;
 	display: block;
 }